+++ R/iotable_get.R
@@ -72,7 +72,8 @@ iotable_get <- function ( labelled_io_data = NULL,
   }
 ##Veryfing source parameter and loading the labelling  ----
   prod_ind <- c("naio_10_cp1700", "naio_10_cp1750", "naio_10_pyp1700",
-                "naio_10_pyp1750", "naio_10_cp1620", "naio_10_cp1630", 
+                "naio_10_pyp1750",
+                "naio_10_cp1610", "naio_10_cp1620", "naio_10_cp1630", 
                 "naio_10_pyp1620", "naio_10_pyp1630" )
   trow_tcol <-  c(  "croatia_2010_1700", "croatia_2010_1800", "croatia_2010_1900")
   croatia_files <- c( "croatia_2010_1700", "croatia_2010_1800", "croatia_2010_1900")
@@ -269,8 +270,6 @@ if ( ! source %in% c("croatia_2010_1700" , "croatia_2010_1800" , "croatia_2010_1
       dplyr::select ( iotables_col, iotables_row, values ) %>% 
       tidyr::spread ( iotables_col, values )
     
-    nrow (iotable_labelled_w )
-    ncol (iotable_labelled_w )    
   } else if ( labelling == "short" & source %in% prod_ind ) {
     
     iotable_labelled_w <- iotable_labelled %>%