+++ R/iotables_download.R
@@ -15,6 +15,8 @@
 ##'  \item{\code{naio_10_pyp1700}}{ Symmetric input-output table at basic prices (product by product) (previous years prices)}
 ##'  \item{\code{naio_10_cp1750}}{ Symmetric input-output table at basic prices (industry by industry)}
 ##'  \item{\code{naio_10_pyp1750}}{ Symmetric input-output table at basic prices (industry by industry) (previous years prices) }
+##'  \item{\code{naio_10_cp1610}}{ Use table at basic prices }
+##'  \item{\code{naio_10_pyp1610}}{ Use table at basic prices (previous years prices) (naio_10_pyp1610) }
 ##'  \item{\code{naio_10_cp1620}}{ Table of trade and transport margins at basic prices}
 ##'  \item{\code{naio_10_pyp1620}}{ Table of trade and transport margins at previous years' prices}
 ##'  \item{\code{naio_10_cp1630}}{ Table of taxes less subsidies on products at basic prices}
@@ -46,6 +48,7 @@ iotables_download <- function ( source = "naio_10_cp1700",
   
   possible_download_sources <- c( "naio_10_cp1700", "naio_10_cp1750", 
                                   "naio_10_pyp1700", "naio_10_pyp1750",
+                                  "naio_10_cp1610", "naio_10_pyp1610", 
                                   "naio_10_cp1620", "naio_10_pyp1620", 
                                   "naio_10_cp1630", "naio_10_pyp1630" )
   source <- tolower (source)
@@ -89,11 +92,11 @@ iotables_download <- function ( source = "naio_10_cp1700",
     dplyr::left_join (., downloaded_labelled, by = c("rows", "values"))
   #message ("Joined labelled and not labelled data.")
   
-  if ( "stk_flow" %in% names ( downloaded )) {
-    downloaded <- downloaded %>%
-      dplyr::filter ( stk_flow == stk_flow )
+  #if ( "stk_flow" %in% names ( downloaded )) {
+  #  downloaded <- downloaded %>%
+  #    dplyr::filter ( stk_flow == stk_flow )
     #message ("Type " , stk_flow, " is returned.")
-  }
+  #}
   
   if ( source == "naio_cp17_r2" ){